% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get Data}
\usage{
get_data(x, return_as_tibble = TRUE, ...)
}
\arguments{
\item{x}{API resource to call.
Options are \code{"books"}, \code{"characters"}, or \code{"houses"}.}

\item{return_as_tibble}{When \code{TRUE}, the output will be a tibble (default); otherwise,
a list will be returned.}

\item{...}{Any other optional \href{https://anapioficeandfire.com/Documentation}{parameter}
that can be provided.}
}
\value{
An object of data for the resource parameter that was requested.
}
\description{
The \code{get_data()} function is used to get structured data from the
\href{https://anapioficeandfire.com/Documentation}{"An API of Ice And Fire"} API service.
}
\details{
By default, only 10 results are returned. This value can be changed using the optional
\code{"pageSize"} parameter. Additionally, to specify which page to access, the optional
\code{"page"} parameter can be used.
}
\examples{
\dontrun{
all_books <- get_data("books", pageSize = 20)
houses_5 <- get_data("houses", page = 2, pageSize = 5)
characters_10 <- get_data("characters")
}
}
